Fix dependency inclusions + Add comments
Also move GoogleAutoToken into /lib.
